chore: Extract buildin methods to FunctionBuilder trait.
